refactor: split components inside main pkg
this commit separates the code a bit and splits the PolicyCreator into a
generic interface
