+++ cmd/permission-generator/defaultmapper.yml
