+++ cmd/permission-generator/policycreator.go
@@ -1,0 +1,5 @@
+package main
+
+type PolicyCreator interface {
+	GeneratePolicy(FuncInvocations) ([]byte, error)
+}
